requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
absl-py==0.15.0
alabaster==0.7.12
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
astunparse==1.6.3
attrs==21.4.0
babel==2.9.1
backcall==0.2.0
bleach==4.1.0
ca-certificates==2021.10.26
cached-property==1.5.2
cachetools==4.2.4
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.10
colorama==0.4.4
cycler==0.11.0
cython==0.29.26
debugpy==1.5.1
decorator==5.1.1
defusedxml==0.7.1
docutils==0.17.1
entrypoints==0.3
flatbuffers==1.12
fonttools==4.28.5
gast==0.3.3
git==2.34.1
google-auth==2.3.3
google-auth-oauthlib==0.4.6
google-pasta==0.2.0
grpcio==1.32.0
h5py==2.10.0
idna==3.3
imageio==2.13.5
imagesize==1.3.0
imgaug==0.4.0
importlib-metadata==4.10.0
importlib-resources==5.4.0
ipykernel==6.7.0
ipyparallel==8.1.0
ipython==7.31.0
ipython-genutils==0.2.0
ipywidgets==7.6.5
jedi==0.18.1
jinja2==3.0.3
joblib==1.1.0
jsonschema==4.4.0
jupyter-client==7.1.0
jupyter-core==4.9.1
jupyterlab-pygments==0.1.2
jupyterlab-widgets==1.0.2
keras==2.7.0
keras-preprocessing==1.1.2
kiwisolver==1.3.2
libclang==12.0.0
markdown==3.3.6
markupsafe==2.0.1
matplotlib==3.5.1
matplotlib-inline==0.1.3
mistune==0.8.4
nbclient==0.5.10
nbconvert==6.4.0
nbformat==5.1.3
nest-asyncio==1.5.4
networkx==2.6.3
nose==1.3.7
notebook==6.4.7
numpy==1.19.5
oauthlib==3.1.1
opencv-python==4.5.5.62
openssl==1.1.1m
opt-einsum==3.3.0
packaging==21.3
pandas==1.3.5
pandocfilters==1.5.0
parso==0.8.3
pickleshare==0.7.5
pillow==9.0.0
pip==21.2.4
prometheus-client==0.12.0
prompt-toolkit==3.0.24
protobuf==3.19.3
psutil==5.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.21
pygments==2.11.2
pyparsing==3.0.6
pyrsistent==0.18.0
python==3.7.11
python-dateutil==2.8.2
pytz==2021.3
pywavelets==1.2.0
pywin32==303
pywinpty==1.1.6
pyzmq==22.3.0
qtconsole==5.2.2
qtpy==2.0.0
requests==2.27.1
requests-oauthlib==1.3.0
rsa==4.8
scikit-image==0.19.1
scikit-learn==1.0.2
scipy==1.7.3
send2trash==1.8.0
setuptools==58.0.4
shapely==1.8.0
six==1.15.0
snowballstemmer==2.2.0
sphinx==4.3.2
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==2.0.0
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.5
sqlite==3.37.0
tensorboard==2.7.0
tensorboard-data-server==0.6.1
tensorboard-plugin-wit==1.8.1
tensorflow==2.4.1
tensorflow-estimator==2.4.0
tensorflow-gpu==2.4.1
tensorflow-io-gcs-filesystem==0.23.1
termcolor==1.1.0
terminado==0.12.1
testpath==0.5.0
threadpoolctl==3.0.0
tifffile==2021.11.2
tornado==6.1
tqdm==4.62.3
traitlets==5.1.1
typing-extensions==3.7.4.3
urllib3==1.26.8
vc==14.2==h21ff451_1
vs2015_runtime==14.27.29016
wcwidth==0.2.5
webencodings==0.5.1
werkzeug==2.0.2
wheel==0.37.1
widgetsnbextension==3.5.2
wincertstore==0.2
wrapt==1.12.1
zipp==3.7.0
